web.2/requirements2.txt
pillow
absl-py==0.6.1
aniso8601==4.0.1
astor==0.7.1
bcrypt==3.1.5
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-JWT-Extended==3.14.0
flask-mongoengine==0.9.5
Flask-PyMongo==2.2.0
Flask-RESTful==0.3.6
Flask-SocketIO==3.1.1
Flask-WTF==0.14.2
gast==0.2.0
grpcio==1.17.1
h5py==2.8.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
MarkupSafe==1.1.0
mongoengine==0.16.3
numpy==1.15.4
protobuf==3.6.1
pycparser==2.19
PyJWT==1.7.1
pymongo==3.7.2
python-engineio==3.1.0
python-socketio==2.1.2
pytz==2018.7
requests==2.21.0
six==1.12.0
tensorboard==1.12.1
tensorflow==1.12.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
Wikipedia-API==0.3.7
WTForms==2.2.1
